Refactor main function into run command
