+++ cli/cmds/root.go
@@ -1,0 +1,5 @@
+package cmds
+
+import "github.com/spf13/cobra"
+
+var Root = &cobra.Command{}
